requirements.txt
altair
numpy
pandas
pydeck
streamlit
pytesseract
mysql-connector-python